Rename: SpacQuizBox 폴더 이동 및 import 경로 수정
