+++ app/(route)/quiz/(components)/SpacQuizBox.tsx
