+++ app/(route)/quiz/page.tsx
@@ -1,0 +1,3 @@
+export default function pages() {
+  return <div>pages</div>;
+}
